Revert "[flutter_tools] post process the gradle log output (#71499)" (#71527)
This reverts commit d688b52a93bc50f784fec189661dca8c2efb6e00.
