+++ packages/flutter_tools/lib/src/android/gradle.dart
@@ -23,7 +23,6 @@ import '../globals.dart' as globals;
 import '../project.dart';
 import '../reporting/reporting.dart';
 import 'gradle_errors.dart';
-import 'gradle_log_processor.dart';
 import 'gradle_utils.dart';
 
 /// The directory where the APK artifact is generated.
@@ -361,7 +360,31 @@ Future<void> buildGradleApp({
   }
   command.add(assembleTask);
 
-  final GradleLogProcessor gradleLogProcessor = GradleLogProcessor(localGradleErrors, globals.logger.isVerbose);
+  GradleHandledError detectedGradleError;
+  String detectedGradleErrorLine;
+  String consumeLog(String line) {
+    // This message was removed from first-party plugins,
+    // but older plugin versions still display this message.
+    if (androidXPluginWarningRegex.hasMatch(line)) {
+      // Don't pipe.
+      return null;
+    }
+    if (detectedGradleError != null) {
+      // Pipe stdout/stderr from Gradle.
+      return line;
+    }
+    for (final GradleHandledError gradleError in localGradleErrors) {
+      if (gradleError.test(line)) {
+        detectedGradleErrorLine = line;
+        detectedGradleError = gradleError;
+        // The first error match wins.
+        break;
+      }
+    }
+    // Pipe stdout/stderr from Gradle.
+    return line;
+  }
+
   final Stopwatch sw = Stopwatch()..start();
   int exitCode = 1;
   try {
@@ -370,14 +393,13 @@ Future<void> buildGradleApp({
       workingDirectory: project.android.hostAppGradleRoot.path,
       allowReentrantFlutter: true,
       environment: gradleEnvironment,
-      mapFunction: gradleLogProcessor.consumeLog,
+      mapFunction: consumeLog,
     );
   } on ProcessException catch (exception) {
-    gradleLogProcessor.atFailureFooter = false;
-    gradleLogProcessor.consumeLog(exception.toString());
+    consumeLog(exception.toString());
     // Rethrow the exception if the error isn't handled by any of the
     // `localGradleErrors`.
-    if (gradleLogProcessor.detectedGradleError == null) {
+    if (detectedGradleError == null) {
       rethrow;
     }
   } finally {
@@ -387,22 +409,22 @@ Future<void> buildGradleApp({
   globals.flutterUsage.sendTiming('build', 'gradle', sw.elapsed);
 
   if (exitCode != 0) {
-    if (gradleLogProcessor.detectedGradleError == null) {
+    if (detectedGradleError == null) {
       BuildEvent('gradle-unknown-failure', flutterUsage: globals.flutterUsage).send();
       throwToolExit(
         'Gradle task $assembleTask failed with exit code $exitCode',
         exitCode: exitCode,
       );
     } else {
-      final GradleBuildStatus status = await gradleLogProcessor.detectedGradleError.handler(
-        line: gradleLogProcessor.detectedGradleErrorLine,
+      final GradleBuildStatus status = await detectedGradleError.handler(
+        line: detectedGradleErrorLine,
         project: project,
         usesAndroidX: usesAndroidX,
         shouldBuildPluginAsAar: shouldBuildPluginAsAar,
       );
 
       if (retries >= 1) {
-        final String successEventLabel = 'gradle-${gradleLogProcessor.detectedGradleError.eventLabel}-success';
+        final String successEventLabel = 'gradle-${detectedGradleError.eventLabel}-success';
         switch (status) {
           case GradleBuildStatus.retry:
             await buildGradleApp(
@@ -432,7 +454,7 @@ Future<void> buildGradleApp({
             // noop.
         }
       }
-      BuildEvent('gradle-${gradleLogProcessor.detectedGradleError.eventLabel}-failure', flutterUsage: globals.flutterUsage).send();
+      BuildEvent('gradle-${detectedGradleError.eventLabel}-failure', flutterUsage: globals.flutterUsage).send();
       throwToolExit(
         'Gradle task $assembleTask failed with exit code $exitCode',
         exitCode: exitCode,
